Escalate the Linux lane to four vCPUs and measure its memory
The instrumented `--all-features --all-targets` build died on
`ubicloud-standard-2` with no diagnostic: `Publish dry run` produced no output
for fifteen minutes and the job then ended, well inside its ninety-minute
timeout and with 31 GB of disk free. That signature is a lost runner rather
than a test failure, and it matches Netsuke's loss under the same workload
shape.

Move the Linux lane to `ubicloud-standard-4` on the recipe's escalation path,
derive `CARGO_BUILD_JOBS` and `NEXTEST_TEST_THREADS` from the lane constant now
set to 4, and register the label with actionlint.

The recipe requires measured evidence before paying for the larger shape, and
none exists because the job died silently. Every Linux job therefore samples
resident memory every fifteen seconds from just after checkout and reports the
peak to the job summary with `if: always()`, so the next run yields the
measurement whether or not it survives. Document the return condition: go back
to `ubicloud-standard-2`, and the constant to 2, once the peak sits well under
6 GB across at least three runs.

The escalation changes the emitted check context to
`build-test (ubicloud-standard-4, stable, true, true, true, true)`.

The resource contracts move to `runner_resources_test.py`, because the
placement module crossed the 400-line limit again.
