+++ .github/actionlint.yaml
@@ -1,3 +1,3 @@
 self-hosted-runner:
   labels:
-    - ubicloud-standard-2
+    - ubicloud-standard-4
